Add authentication functionality and related packages
Integrated `next-auth` package for user authentication in the application. Added the usage of `authenticate` function for handling user logins as well as the relevant dependencies in `pnpm-lock.yaml` and `package.json` files. Created `auth.ts` and `auth.config.ts` files to handle the authentication configuration and created a new login form along with related page and middleware for handling user logins. Also enhanced the side navigation with a sign-out feature. The changes are intended to enhance the user experience by providing a secure login feature.