Refactor code structure and improve readability
This commit is revolving around code refactor for improved readability and structure. The 'auth.ts', 'middleware.ts' and 'auth.config.ts' files are now at root level for better accessibility. The path for importing signIn function in 'actions.ts' is changed accordingly. In 'auth.ts', getUser function's SQL query format is cleaned up to enhance readability. In 'middleware.ts', comments are added for 'matcher' explaining what it does. Layout changes with the intent to improve clarity of code are made in 'login-form.tsx'. Import paths for AcmeLogo and signOut are updated in 'sidenav.tsx' according to the new file locations. Lastly, some minor changes to sort the code and add explanatory comments are made in 'auth.config.ts'.